move Knock policies to the CommonwealthConsumer
