+++ libs/model/src/policies/NotificationsSettings.policy.ts
@@ -1,12 +1,11 @@
 import { Policy } from '@hicommonwealth/core';
 import { events } from '@hicommonwealth/schemas';
-import { processSubscriptionPreferencesUpdated } from './subscriptionPreferencesUpdated';
+import { processSubscriptionPreferencesUpdated } from '../services/knock/subscriptionPreferencesUpdated';
 
 const notificationSettingsInputs = {
   SubscriptionPreferencesUpdated: events.SubscriptionPreferencesUpdated,
 };
 
-// TODO: this should be in libs/model, but currently depending on libs/adapter for config
 export function NotificationsSettingsPolicy(): Policy<
   typeof notificationSettingsInputs
 > {
