+++ libs/model/src/policies/index.ts
@@ -8,5 +8,6 @@ export * from './FarcasterWorker.policy';
 export * from './Launchpad.policy';
 export * from './Nominations.policy';
 export * from './Notifications.policy';
+export * from './NotificationsSettings.policy';
 export * from './ReactionWorker.policy';
 export * from './TwitterEngagement.policy';
