+++ packages/commonwealth/package.json
@@ -107,7 +107,7 @@
     "@keplr-wallet/types": "^0.12.23",
     "@keplr-wallet/unit": "^0.12.23",
     "@knocklabs/client": "^0.10.16",
-    "@knocklabs/node": "^0.6.13",
+    "@knocklabs/node": "catalog:",
     "@knocklabs/react": "^0.2.32",
     "@knocklabs/react-notification-feed": "^0.8.15",
     "@lexical/rich-text": "^0.17.0",
