+++ packages/commonwealth/server/bindings/rascalConsumerMap.ts
@@ -17,11 +17,11 @@ import {
   LaunchpadPolicy,
   NominationsWorker,
   NotificationsPolicy,
+  NotificationsSettingsPolicy,
   ReactionWorker,
   TwitterEngagementPolicy,
   User,
 } from '@hicommonwealth/model';
-import { NotificationsSettingsPolicy } from '../workers/knock/NotificationsSettings.policy';
 
 const _ContestWorker: Consumer<ReturnType<typeof ContestWorker>> = {
   consumer: ContestWorker,
@@ -57,12 +57,10 @@ const _Xp: Consumer<ReturnType<typeof User.Xp>> = {
 
 const _NotificationsSettingsPolicy = {
   consumer: NotificationsSettingsPolicy,
-  worker: 'knock',
 };
 
 const _NotificationsPolicy = {
   consumer: NotificationsPolicy,
-  worker: 'knock',
   // This disables retry strategies on any handler error/failure
   // This is because we cannot guarantee whether a Knock workflow trigger
   // call was successful or not. It is better to 'miss' notifications then
