+++ pnpm-workspace.yaml
@@ -2,3 +2,6 @@ packages:
   - 'libs/**'
   - 'packages/commonwealth'
   - 'packages/load-testing'
+
+catalog:
+  '@knocklabs/node': ^0.6.13
